Extract active materialized query execution
